Changes minimum version from 1809 to 1803
We are lowering the version is since everyone has not yet gotten the upgrade from version 1803 to 1809, and the next May update requires 15GB extra reserved space on the harddrive, which may cause several computers to not get the update due do insufficient disk space.

Removed IsEditable= "True" from the LidarPages ComboBox (since that requires 1809 to work).

The Shutdown and Restart functions in the PowerViewModel also requires 1809, but since the buttons activating these functions are only available when running on the vehicle, and hidden when running on PC (which may have 1803), these functions have been left in the program.