Changes from Task to void and removed Task.Delay
Changed from async Task to async void since this method was not awiated, and there is a Try-Catch inside the method.

The task.delay was only being utilzed for cancelling the task.
Removed the option to set a delay, and replaced the loopcheck from beeing infinite to checking if cancellation has been requested.

Also removed variable from catch block since it was not beeing used.

Note that this method is not considered fire-and forget, since it can be cancelled.