Adds attribute to fix documentation statistics
This comitt modifies .gitattributes so that the documents added to [Drawings and Documents] folder should not be included in the GitHub statistics.

This is nessecarry since adding the 3D moddel of the car caused many lines to be added to the code statistics:
- 85 313 lines for the .iges file.
- 74 885 lines for the .sat file.
- 78 015 lines for the .smt file.
- 69 599 lines for the .step file.

This casues the statistics page to over-report by 307 919 additions.

With this commit, those lines should no longer be counted as it gets configured to be interpreted as 'documentation'