Makes simulator calculate realistic encoder data
The encoder data when driving in a straight line is accurate.

The encoder data when car is rotating gives an approximate accuracy. The encoder data when the real car is turning can't be trusted since it relies on skid-steering. For that reason, we have also added this as a factor in the simulator to encurage the making of control logic which discards the encoder data collected while rotating.