Modifies the Control Logic DriveToLargestDistance
Increases the distance at front required before it stops rotating (on the spot) away from obstruction to 1.3 meters.

Also makes it increase speed gradually if it is unable to rotate away from obstruction (i.e. requested power to wheels not enough to cause it to rotate on the spot).

Also adds information in the description text regarding startup time for lidar.