Adds shutdown and restart to extras app
The target platform version was increased to 1809 to allow the shutdown and restart app functions to work.
Also adds SystemManagement capabilities in package manifest, since this is required to use the shutdown and restart operations.

The functions for shutdown and restart were allready written prior to this commit. But this commit unncomments the lines for the code so that it becomes functional.