Removes resolved TODO for simulated encoder
The linear displacement inaccuraccy was solved by storing the difference between the truncated distance (to be sent to main program) and the float value, back into the CmTraveled.. variable.

See "_vehicle.CmTraveledLeftWheel = _vehicle.CmTraveledLeftWheel - distance;"
in CmSinceLastRequest() method of the SimulatedEncoderSensor class.

The comment is therefore no longer valid, and is thus removed in this commit.