Adds escepe character to the brackets
The lines that previously were added did not succeed in ignoring the documents folder from the statistics count, even after waiting for 13 days (the statistics is caluclated as a background task).

Upon the suggestion of GitHub support, this commit adds escape character in front of the brackets in the folder name, as these may be causing the issue.